+++ src/com/koing/two/Movie.java
@@ -1,6 +1,6 @@
 package com.koing.two;
 
-import com.koing.two.discountPolicy.DiscountPolicy;
+import com.koing.two.discountPolicy.DefaultDiscountPolicy;
 
 import java.time.Duration;
 
@@ -8,9 +8,9 @@ public class Movie {
     private String title;
     private Duration runningTime;
     private Money fee;
-    private DiscountPolicy discountPolicy;
+    private DefaultDiscountPolicy discountPolicy;
 
-    public Movie(String title, Duration runningTime, Money fee, DiscountPolicy discountPolicy) {
+    public Movie(String title, Duration runningTime, Money fee, DefaultDiscountPolicy discountPolicy) {
         this.title = title;
         this.runningTime = runningTime;
         this.fee = fee;
