+++ src/com/koing/two/discountPolicy/DiscountPolicy.java
@@ -2,27 +2,7 @@ package com.koing.two.discountPolicy;
 
 import com.koing.two.Money;
 import com.koing.two.Screening;
-import com.koing.two.discountCondition.DiscountCondition;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-
-public abstract class DiscountPolicy {
-    private List<DiscountCondition> conditions = new ArrayList<>();
-
-    public DiscountPolicy(DiscountCondition ... conditions) {
-        this.conditions = Arrays.asList(conditions);
-    }
-
-    public Money calculateDiscountAmount(Screening screening) {
-        for(DiscountCondition each : conditions) {
-            if (each.isSatisfiedBy(screening)) {
-                return getDiscountAmount(screening);
-            }
-        }
-        return Money.ZERO;
-    }
-
-    abstract protected Money getDiscountAmount(Screening screening);
+public interface DiscountPolicy {
+    Money calculateDiscountAmount(Screening screening);
 }
