+++ src/com/koing/two/discountPolicy/NoneDiscountPolicy.java
@@ -3,9 +3,10 @@ package com.koing.two.discountPolicy;
 import com.koing.two.Money;
 import com.koing.two.Screening;
 
-public class NoneDiscountPolicy extends DiscountPolicy {
+public class NoneDiscountPolicy implements DiscountPolicy {
+
     @Override
-    protected Money getDiscountAmount(Screening screening) {
+    public Money calculateDiscountAmount(Screening screening) {
         return Money.ZERO;
     }
 }
